forensic_timeline/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
elasticsearch==7.17.0
pandas==1.5.3
python-dotenv==1.0.0
pytest==7.3.1
requests==2.28.2
jupyter==1.0.0
scikit-learn==1.2.2
joblib==1.2.0
